+++ HISTORY.rst
@@ -3,6 +3,13 @@ History
 -------
 
 
+1.1.0 (2014-03-21)
+++++++++++++++++++
+
+- Support Windows operating system
+- Download and unzip wakatime-cli first time plugin is run
+
+
 1.0.0 (2014-03-21)
 ++++++++++++++++++